requirements.txt
django>=4.0.2,<4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.11.0
psycopg2-binary==2.8.6
python-dotenv==0.19.1